updates
I2C update, LUA added ( 5.2, sources a bit later ), USB WiFI added
RT2501 and RT2870 chip set,

+++ SAMPLES/DSHOWFILTERS/mpeg4ip_mp4v2/include/mp4property.h
@@ -441,7 +441,7 @@ public:
     }
 
     u_int32_t GetCount() {
-      return m_pCountProperty->GetValue();
+      return (u_int32_t)m_pCountProperty->GetValue();
     }
     void SetCount(u_int32_t count) {
       m_pCountProperty->SetValue(count);
